refactor: move code around for emoji plugin
